feat: update login, add layou theme and update e2e flows
