+++ lib/ui/views/login/login_view.dart
@@ -1,125 +1,143 @@
-// This file is not being used, it should be deleted soon
-
-import 'package:boleia_app/ui/views/halpers/functions.dart';
-import 'package:boleia_app/ui/views/passagers/passager_signup/signup_page.dart';
-import 'package:boleia_app/ui/views/start/start_page.dart';
-import 'package:boleia_app/ui/widgets/button.dart';
-import 'package:boleia_app/ui/widgets/formfield.dart';
+import 'package:boleia_app/ui/widgets/login_button.dart';
+import 'package:boleia_app/ui/widgets/square_tile.dart';
 import 'package:flutter/material.dart';
+import 'package:boleia_app/ui/widgets/login_text_field.dart';
+import 'package:boleia_app/ui/views/halpers/functions.dart';
+import 'package:boleia_app/ui/views/start/start_page.dart';
+import 'package:boleia_app/ui/views/passagers/passager_signup/signup_page.dart';
 
 class LoginPage extends StatelessWidget {
-  const LoginPage({super.key});
+  LoginPage({super.key});
+
+  // text editing controllers
+  final emailController = TextEditingController();
+
+  final passwordController = TextEditingController();
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: SingleChildScrollView(
+      backgroundColor: Colors.grey[300],
+      // SafeArea creates a Widget that avoids the notch area, and the other operating system interfaces (such as Edge borders, etc)
+      body: SafeArea(
         child: Center(
-          child: Padding(
-            padding: const EdgeInsets.all(16.0),
+          child: SingleChildScrollView(
             child: Column(
+              mainAxisAlignment: MainAxisAlignment.center,
               children: [
-                const SizedBox(height: 80),
-                const Text(
-                  "Login",
+                const SizedBox(height: 50),
+                // logo
+                const Icon(
+                  Icons.car_crash_rounded,
+                  size: 100,
+                ),
+                const SizedBox(height: 50),
+                // welcome back, you've been missed
+                Text(
+                  'Bem vindo de volta, sentimos sua falta',
                   style: TextStyle(
-                      color: Colors.blue,
-                      fontSize: 40,
-                      fontWeight: FontWeight.bold),
-                ),
-                const SizedBox(height: 20),
-                SizedBox(
-                  child: AppButton(
-                    color: Colors.red,
-                    fcolor: Colors.white,
-                    borderColor: Colors.red,
-                    icon: Icons.facebook,
-                    hasIcon: true,
-                    label: "Entrar com o Google",
-                    height: 60.0,
-                    width: double.infinity,
-                    onPressed: () {
-                      navegateReplaceTo(context, const StartPage());
-                    },
+                    color: Colors.grey[700],
+                    fontSize: 16,
                   ),
                 ),
-                const SizedBox(height: 20),
-                SizedBox(
-                  child: AppButton(
-                    color: Colors.blue,
-                    fcolor: Colors.white,
-                    borderColor: Colors.blue,
-                    icon: Icons.facebook,
-                    hasIcon: true,
-                    label: "Entrar com o Facebook",
-                    height: 60.0,
-                    width: double.infinity,
-                    onPressed: () {
-                      navegateReplaceTo(context, const StartPage());
-                    },
-                  ),
+                const SizedBox(height: 25),
+
+                // username textfield
+                LoginTextField(
+                  controller: emailController,
+                  hintText: 'Email',
+                  obscureText: false,
                 ),
-                const SizedBox(height: 20),
-                const Text(
-                  "_____________ OR ______________",
-                  style: TextStyle(color: Colors.grey),
-                ),
-                const SizedBox(height: 20),
-                Formfield(
-                  callbackFunction: (x) {},
-                  hintText: "Digite o seu email",
-                  type: "TEXT",
-                  isEnabled: true,
-                  isPassword: false,
-                  nameController: TextEditingController(),
-                ),
-                const SizedBox(height: 20),
-                Formfield(
-                  callbackFunction: (x) {},
-                  hintText: "Digite a sua senha",
-                  type: "TEXT",
-                  isEnabled: true,
-                  isPassword: true,
-                  nameController: TextEditingController(),
-                ),
-                const SizedBox(height: 40),
-                SizedBox(
-                  width: double.infinity,
-                  child: AppButton(
-                    color: Colors.green,
-                    fcolor: Colors.white,
-                    borderColor: Colors.green,
-                    hasIcon: false,
-                    label: "Login",
-                    height: 60.0,
-                    width: double.infinity,
-                    onPressed: () {
-                      navegateReplaceTo(context, const StartPage());
-                    },
-                  ),
-                ),
+
                 const SizedBox(height: 10),
-                SizedBox(
-                  width: double.infinity,
-                  child: Center(
-                    child: Row(
-                      crossAxisAlignment: CrossAxisAlignment.center,
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: [
-                        const Text("Ainda nao possue uma conta? "),
-                        InkWell(
-                          child: const Text(
-                            "Crie uma!",
-                            style: TextStyle(color: Colors.blue),
-                          ),
-                          onTap: () {
-                            navegateTo(context, const SignupPage());
-                          },
-                        ),
-                      ],
-                    ),
+
+                // password textfield
+                LoginTextField(
+                  controller: passwordController,
+                  hintText: 'Senha',
+                  obscureText: true,
+                ),
+
+                const SizedBox(height: 10),
+                // forgot password?
+                Padding(
+                  padding: const EdgeInsets.symmetric(horizontal: 25),
+                  child: Row(
+                    mainAxisAlignment: MainAxisAlignment.end,
+                    children: [
+                      Text('Esqueceu sua senha?',
+                          style: TextStyle(color: Colors.grey[600])),
+                    ],
                   ),
+                ),
+                const SizedBox(height: 25),
+                // sign in button
+                LoginButton(
+                  onTap: () {
+                    navegateReplaceTo(context, const StartPage());
+                  },
+                  text: 'Iniciar sessão',
+                ),
+
+                const SizedBox(height: 50),
+
+                // or continue with
+                Row(
+                  children: [
+                    Expanded(
+                      child: Divider(
+                        thickness: 0.5,
+                        color: Colors.grey[400],
+                      ),
+                    ),
+                    Text(
+                      'Ou continuar com',
+                      style: TextStyle(color: Colors.grey[700]),
+                    ),
+                    Expanded(
+                      child: Divider(
+                        thickness: 0.5,
+                        color: Colors.grey[400],
+                      ),
+                    ),
+                  ],
+                ),
+                const SizedBox(height: 50),
+                const Row(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    SquareTile(imagePath: 'assets/google.png'),
+                    SizedBox(width: 25),
+                    SquareTile(imagePath: 'assets/facebook.png')
+                  ],
+                ),
+                const SizedBox(height: 50),
+                Row(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    Text(
+                      'Não tem uma conta?',
+                      style: TextStyle(
+                        color: Colors.grey[700],
+                      ),
+                    ),
+                    const SizedBox(width: 4),
+                    GestureDetector(
+                      onTap: () {
+                        navegateTo(context, RegisterPage());
+                      },
+                      child: const Text(
+                        'Criar uma agora',
+                        style: TextStyle(
+                            color: Colors.blue, fontWeight: FontWeight.bold),
+                      ),
+                    ),
+                  ],
                 )
+
+                // google + apple sign in buttons
+
+                // not a member? register now
               ],
             ),
           ),
