+++ lib/ui/views/passagers/passager_signup/signup_page.dart
@@ -1,124 +1,137 @@
-// This file is not being used, it should be deleted soon
-
-import 'package:boleia_app/ui/views/halpers/functions.dart';
-import 'package:boleia_app/ui/views/passagers/home/home_page.dart';
-import 'package:boleia_app/ui/widgets/button.dart';
-import 'package:boleia_app/ui/widgets/formfield.dart';
+import 'package:boleia_app/ui/views/login/login_view.dart';
+import 'package:boleia_app/ui/widgets/login_text_field.dart';
+import 'package:boleia_app/ui/widgets/square_tile.dart';
 import 'package:flutter/material.dart';
+import 'package:boleia_app/ui/widgets/login_button.dart';
+import 'package:boleia_app/ui/views/halpers/functions.dart';
 
-class SignupPage extends StatefulWidget {
-  const SignupPage({super.key});
+class RegisterPage extends StatelessWidget {
+  RegisterPage({super.key});
 
-  @override
-  State<SignupPage> createState() => _SignupPageState();
-}
+  // text editing controllers
+  final emailController = TextEditingController();
 
-class _SignupPageState extends State<SignupPage> {
-  final List<String> parentItems = [
-    'Item1',
-    'Item2',
-    'Item3',
-    'Item4',
-  ];
+  final passwordController = TextEditingController();
+  final confirmPasswordController = TextEditingController();
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: Text("Registar Passageiro")),
-      body: SingleChildScrollView(
-        child: Padding(
-          padding: const EdgeInsets.all(16.0),
-          child: Column(
-            children: [
-              Formfield(
-                callbackFunction: (x) {},
-                hintText: "Nome completo",
-                type: "TEXT",
-                isEnabled: true,
-                isPassword: false,
-                nameController: TextEditingController(),
-              ),
-              const SizedBox(height: 10),
-              Formfield(
-                callbackFunction: (x) {},
-                hintText: "Endereco",
-                type: "TEXT",
-                isEnabled: true,
-                isPassword: false,
-                nameController: TextEditingController(),
-              ),
-              const SizedBox(height: 10),
-              Formfield(
-                callbackFunction: (x) {},
-                hintText: "Telefone",
-                type: "TEL",
-                isEnabled: true,
-                isPassword: false,
-                nameController: TextEditingController(),
-              ),
-              const SizedBox(height: 10),
-              Formfield(
-                callbackFunction: (x) {},
-                hintText: "Telefone  de emergência",
-                type: "TEL",
-                isEnabled: true,
-                isPassword: false,
-                nameController: TextEditingController(),
-              ),
-              const SizedBox(height: 10),
-              Formfield(
-                callbackFunction: (x) {},
-                hintText: "Telefone  de emergência",
-                type: "TEL",
-                isEnabled: true,
-                isPassword: false,
-                nameController: TextEditingController(),
-              ),
-              const SizedBox(height: 10),
-              Formfield(
-                callbackFunction: (x) {},
-                hintText: "Grau de parentesco",
-                type: "SELECT",
-                items: parentItems,
-                isEnabled: true,
-                isPassword: false,
-                nameController: TextEditingController(),
-              ),
-              const SizedBox(height: 10),
-              Formfield(
-                callbackFunction: (x) {},
-                hintText: "Endereco de origem",
-                type: "TEXT",
-                isEnabled: true,
-                isPassword: false,
-                nameController: TextEditingController(),
-              ),
-              const SizedBox(height: 10),
-              Formfield(
-                callbackFunction: (x) {},
-                hintText: "Endereco de destino",
-                type: "TEXT",
-                isEnabled: true,
-                isPassword: false,
-                nameController: TextEditingController(),
-              ),
-              const SizedBox(height: 10),
-              SizedBox(
-                width: double.infinity,
-                child: AppButton(
-                  color: Colors.green,
-                  fcolor: Colors.white,
-                  borderColor: Colors.green,
-                  hasIcon: false,
-                  label: "Login",
-                  height: 60.0,
-                  width: double.infinity,
-                  onPressed: () {
-                    navegateReplaceTo(context, const HomePage());
-                  },
+      backgroundColor: Colors.grey[300],
+      // SafeArea creates a Widget that avoids the notch area, and the other operating system interfaces (such as Edge borders, etc)
+      body: SafeArea(
+        child: Center(
+          child: SingleChildScrollView(
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                const SizedBox(height: 25),
+                // logo
+                const Icon(
+                  Icons.car_crash_rounded,
+                  size: 65,
                 ),
-              ),
-            ],
+                const SizedBox(height: 25),
+                // welcome back, you've been missed
+                Text(
+                  'Olá, vamos cria uma nova conta',
+                  style: TextStyle(
+                    color: Colors.grey[700],
+                    fontSize: 16,
+                  ),
+                ),
+                const SizedBox(height: 25),
+
+                // username textfield
+                LoginTextField(
+                  controller: emailController,
+                  hintText: 'Email',
+                  obscureText: false,
+                ),
+
+                const SizedBox(height: 10),
+
+                // password textfield
+                LoginTextField(
+                  controller: passwordController,
+                  hintText: 'Senha',
+                  obscureText: true,
+                ),
+                const SizedBox(height: 10),
+                LoginTextField(
+                  controller: confirmPasswordController,
+                  hintText: 'Confirmar senha',
+                  obscureText: true,
+                ),
+
+                const SizedBox(height: 25),
+                // sign in button
+                LoginButton(
+                  onTap: () {},
+                  text: 'Criar conta',
+                ),
+
+                const SizedBox(height: 50),
+
+                // or continue with
+                Row(
+                  children: [
+                    Expanded(
+                      child: Divider(
+                        thickness: 0.5,
+                        color: Colors.grey[400],
+                      ),
+                    ),
+                    Text(
+                      'Ou criar com',
+                      style: TextStyle(color: Colors.grey[700]),
+                    ),
+                    Expanded(
+                      child: Divider(
+                        thickness: 0.5,
+                        color: Colors.grey[400],
+                      ),
+                    ),
+                  ],
+                ),
+                const SizedBox(height: 50),
+                const Row(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    SquareTile(imagePath: 'assets/google.png'),
+                    SizedBox(width: 25),
+                    SquareTile(imagePath: 'assets/facebook.png')
+                  ],
+                ),
+                const SizedBox(height: 50),
+                Row(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    Text(
+                      'Já tem uma conta?',
+                      style: TextStyle(
+                        color: Colors.grey[700],
+                      ),
+                    ),
+                    const SizedBox(width: 4),
+                    GestureDetector(
+                      onTap: () {
+                        navegateTo(context, LoginPage());
+                      },
+                      child: const Text(
+                        'Entrar agora',
+                        style: TextStyle(
+                            color: Colors.blue, fontWeight: FontWeight.bold),
+                      ),
+                    ),
+                  ],
+                )
+
+                // google + apple sign in buttons
+
+                // not a member? register now
+              ],
+            ),
           ),
         ),
       ),
