+++ lib/ui/views/start/start_page.dart
@@ -1,6 +1,7 @@
 import 'package:boleia_app/ui/views/halpers/functions.dart';
 import 'package:boleia_app/ui/views/passagers/home/home_page.dart';
 import 'package:boleia_app/ui/views/drivers/home_driver/home_driver.dart';
+import 'package:boleia_app/ui/views/login/login_view.dart';
 import 'package:boleia_app/ui/widgets/button.dart';
 import 'package:flutter/material.dart';
 
@@ -10,18 +11,38 @@ class StartPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: Center(
-        child: Padding(
-          padding: const EdgeInsets.all(16.0),
+      backgroundColor: Colors.grey[300],
+      appBar: AppBar(
+        backgroundColor: Colors.grey[300],
+        actions: [
+          GestureDetector(
+              onTap: () {
+                navegateReplaceTo(context, LoginPage());
+              },
+              child: const Padding(
+                padding: EdgeInsets.symmetric(horizontal: 15),
+                child: Text(
+                  'Sign out',
+                  style: TextStyle(
+                    fontWeight: FontWeight.bold,
+                    fontSize: 16,
+                  ),
+                ),
+              ))
+        ],
+      ),
+      body: SafeArea(
+        child: Center(
           child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
             children: [
-              const SizedBox(height: 100),
+              // const SizedBox(height: 100),
               const CircleAvatar(
                 foregroundImage: AssetImage("assets/logo.png"),
-                minRadius: 50,
-                maxRadius: 75,
+                minRadius: 65,
+                maxRadius: 90,
               ),
-              const SizedBox(height: 60),
+              const SizedBox(height: 150),
               SizedBox(
                 width: 250,
                 child: AppButton(
